feat: 판매된 공연권 개수(#548)

+++ src/main/java/PodoeMarket/podoemarket/profile/dto/response/RequestedPerformanceResponseDTO.java
@@ -31,7 +31,7 @@ public class RequestedPerformanceResponseDTO {
         private Integer scriptQuantity;
         private Boolean performance;
         private Long performancePrice;
-        private Integer performanceQuantity;
+        private Long performanceQuantity;
     }
 
     @Data
